+++ ts/features/messages/components/MessageAttachment/__test__/PdfViewer.test.tsx
@@ -2,12 +2,12 @@ import * as React from "react";
 import { render } from "@testing-library/react-native";
 import configureMockStore from "redux-mock-store";
 import { Provider } from "react-redux";
-import { DSPdfViewer } from "../DSPdfViewer";
+import { PdfViewer } from "../PdfViewer";
 import { appReducer } from "../../../../../store/reducers";
 import { applicationChangeState } from "../../../../../store/actions/application";
 import { GlobalState } from "../../../../../store/reducers/types";
 
-describe("DSPdfViwer", () => {
+describe("PdfViewer", () => {
   it("should match the snapshot", () => {
     const globalState = appReducer(undefined, applicationChangeState("active"));
     const enrichedState = {
@@ -21,7 +21,7 @@ describe("DSPdfViwer", () => {
     const store: ReturnType<typeof mockStore> = mockStore(enrichedState);
     const component = render(
       <Provider store={store}>
-        <DSPdfViewer downloadPath={"file:///randomFile.pdf"} />
+        <PdfViewer downloadPath={"file:///randomFile.pdf"} />
       </Provider>
     );
     expect(component.toJSON()).toMatchSnapshot();
